sharedElastest ini

+++ vars/elastest_lib.groovy
@@ -50,10 +50,14 @@ class elastest_lib implements Serializable {
 	*/
 	def setShared( boolean value) { this.@shared = value }
 	
+	
 	/*
 	*	@return elastest_ip
 	*/
-	def getIp() { return this.@ip }
+	def getIp() { 
+		if (this.@shared)
+			return $SHARED_ELASTEST_IP
+	return this.@ip }
 
 	/*
 	*	@return elastest_port
@@ -63,7 +67,11 @@ class elastest_lib implements Serializable {
 	/*
 	*	@ return full conection url for torm
 	*/
-	def getEtmUrl() { return 'http://'+this.@ip+':'+this.@port }
+	def getEtmUrl() { 
+		if (this.@shared)
+				return 'http://'+$SHARED_ELASTEST_IP+':'+this.@port  
+		return 'http://'+this.@ip+':'+this.@port 
+	}
 	
 	/*
 	*	When using methods of this library inside any provided step should be called through the 
@@ -72,19 +80,35 @@ class elastest_lib implements Serializable {
 	*/
 	def getElastestMethods(){ return this }
 	
+	
+	/*
+	* checks availability of remote elastest
+	*/
+	def testRemoteElastest(){
+		try {  
+				echo '[INI] testRemoteElastest'			
+				new URL('http://'+$SHARED_ELASTEST_IP+':'+this.@port).getText()
+				echo '[END] testRemoteElastest'
+				return 0
+				
+			} catch (Exception e) {
+				echo '[END] testRemoteElastest'
+				return 1
+			}
+		}
 	/*
 	*	Connect container to elastest network
 	*/
 	def connect2ElastestNetwork(){
 		echo '[INI] connect2ElastestNetwork'
-		
-		def containerId= this.@ctx.sh (
-			script: 'cat /proc/self/cgroup | grep "docker" | sed s/\\\\//\\\\n/g | tail -1',
-			returnStdout: true
-		).trim() 
-		echo "containerId = ${containerId}"
-		this.@ctx.sh "docker network connect elastest_elastest "+ containerId
-		
+		if (! this.@shared){
+			def containerId= this.@ctx.sh (
+				script: 'cat /proc/self/cgroup | grep "docker" | sed s/\\\\//\\\\n/g | tail -1',
+				returnStdout: true
+			).trim() 
+			echo "containerId = ${containerId}"
+			this.@ctx.sh "docker network connect elastest_elastest "+ containerId
+		}
 		echo '[END] connect2ElastestNetwork'
 	}
 	
@@ -95,10 +119,14 @@ class elastest_lib implements Serializable {
 	*/
 	def startElastest(){
 		echo '[INI] startElastest'
-		
-		def start_elastest_result = this.@ctx.sh script: 'docker run -d --name="elastest_platform" -v /var/run/docker.sock:/var/run/docker.sock --rm elastest/platform:'+this.@version+' start --pullcore '+ this.@mode, returnStatus:true
-		echo 'startElastest-- start_elastest_result = '+start_elastest_result
-		
+		def start_elastest_result = 1
+		if (! this.@shared){
+			start_elastest_result = this.@ctx.sh script: 'docker run -d --name="elastest_platform" -v /var/run/docker.sock:/var/run/docker.sock --rm elastest/platform:'+this.@version+' start --pullcore '+ this.@mode, returnStatus:true
+			echo 'startElastest-- start_elastest_result = '+start_elastest_result
+		}
+		else {
+			start_elastest_result = testRemoteElastest()
+		}
 		echo '[END] startElastest'
 		return (start_elastest_result==0)
 	}
@@ -109,10 +137,19 @@ class elastest_lib implements Serializable {
 	*/
 	def elastestIsRunning(){
 		echo '[INI] elastestIsRunning'
-		def platform_state = this.@ctx.sh script: 'docker ps | grep elastest_platform | grep -c Up', returnStatus:true
-		def etm_state = this.@ctx.sh script: ' docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' wait --running=0 ', returnStatus:true
+		def elastest_state = false
+		
+		if (this.@shared){
+			def platform_state = this.@ctx.sh script: 'docker ps | grep elastest_platform | grep -c Up', returnStatus:true
+			def etm_state = this.@ctx.sh script: ' docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' wait --running=0 ', returnStatus:true
+			elastest_state = (platform_state==0 && etm_state==0)
+		}
+		else {
+			def remoteElastest = testRemoteElastest()
+			elastest_state = (remoteElastest==0)
+		}
+		return elastest_state
 		echo '[END] elastestIsRunning'
-		return (platform_state==0 && etm_state==0)
 	}
 
 	/*
@@ -120,9 +157,16 @@ class elastest_lib implements Serializable {
 	*	@return boolean
 	*/
 	def waitElastest(){
-		echo '[INI] waitElastest'
-		def elastest_is_running = this.@ctx.sh script: 'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' wait --container=900', returnStatus:true
-		echo '[END] waitElastest'
+		echo '[INI] elastestIsRunning'
+		def elastest_is_running = false
+		
+		if (this.@shared){
+			elastest_is_running = this.@ctx.sh script: 'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' wait --container=900', returnStatus:true
+		}
+		else {
+			elastest_is_running = testRemoteElastest()
+		}
+		
 		return (elastest_is_running == 0)
 	}
 
@@ -131,13 +175,16 @@ class elastest_lib implements Serializable {
 	*/
 	def stopElastest(){
 		echo '[INI] stopElastest'
-		def start_elastest_result = this.@ctx.sh script:'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' stop', returnStatus:true
-        try {
-            this.@ctx.sh "docker rm -f elastest-platform"
-        }catch(e) {
-            echo "Error: $e"
-        }
-		echo 'start_elastest_result = '+start_elastest_result	
+		if (this.@shared){
+			def start_elastest_result = this.@ctx.sh script:'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock elastest/platform:'+this.@version+' stop', returnStatus:true
+			try {
+				this.@ctx.sh "docker rm -f elastest-platform"
+			}catch(e) {
+				echo "Error: $e"
+			}
+			echo 'start_elastest_result = '+start_elastest_result	
+		}
+		
 		echo '[END] stopElastest'
 	}
 	
@@ -146,32 +193,36 @@ class elastest_lib implements Serializable {
 	*/
 	def getApi(){
 		echo '[INI] getAPI'
-		
-		def get_api= this.@ctx.sh (
-                script: 'docker inspect --format=\\"{{.NetworkSettings.Networks.elastest_elastest.Gateway}}\\" elastest_etm-proxy_1',
-                returnStdout: true
-            ).trim()
-        echo "ETM container IP=${get_api}"
-		
-		//this is valid for output:
-		/*
-		Waiting for ETM...
-		Platform inserted into network succesfully
+		if (this.@shared){
+			def get_api= this.@ctx.sh (
+					script: 'docker inspect --format=\\"{{.NetworkSettings.Networks.elastest_elastest.Gateway}}\\" elastest_etm-proxy_1',
+					returnStdout: true
+				).trim()
+			echo "ETM container IP=${get_api}"
+			
+			//this is valid for output:
+			/*
+			Waiting for ETM...
+			Platform inserted into network succesfully
 
-		Container created with IP: 172.18.0.19
-		ETM is ready in http://172.18.0.19:8091
+			Container created with IP: 172.18.0.19
+			ETM is ready in http://172.18.0.19:8091
 
-		ElasTest API info:
-		Url: http://172.18.0.19:8091
-		*/
-		
-		/*def strings = get_api.tokenize( ':' )
-		this.@ip = strings[strings.size()-2].replaceAll("/",'')
-		this.@port = strings[strings.size()-1]
-		*/
-		this.@ip=get_api
-		this.@ip = this.@ip.replaceAll('"','')
-		echo '[END] getAPI  ip:port ' +this.@ip+':'+this.@port
+			ElasTest API info:
+			Url: http://172.18.0.19:8091
+			*/
+			
+			/*def strings = get_api.tokenize( ':' )
+			this.@ip = strings[strings.size()-2].replaceAll("/",'')
+			this.@port = strings[strings.size()-1]
+			*/
+			this.@ip=get_api
+			this.@ip = this.@ip.replaceAll('"','')
+			echo '[END] getAPI  ip:port ' +this.@ip+':'+this.@port
+		}
+		else {
+			echo '[END] getAPI  ip:port ' +$SHARED_ELASTEST_IP+':'+this.@port
+		}
 	}
 	
 	/*
@@ -229,30 +280,9 @@ class elastest_lib implements Serializable {
 			}
 		}
 		else {
-			this.@ctx.node('commonE2E'){
+			this.@ctx.node('docker'){
 				this.@ctx.stage ('launch elastest')
-					echo "sharedElastest ="+this.@shared
-					def elastest_is_running = elastestIsRunning()
-					if (elastest_is_running){ //stop and start again --> elastest is unique and frethis.@ctx.sh with each start
-						stopElastest()
-						sleep(10)//TODO: change for method like waitToStop or something like that
-						elastest_is_running = elastestIsRunning()
-						echo 'elastest_is_running:'+elastest_is_running
-						if (elastest_is_running){
-							this.@ctx.currentBuild.result = 'FAILURE'
-							return
-						}	
-					}
-					startElastest()
-					//lets set more time for waiting
-					def counter = 3
-					while (!elastest_is_running && counter > 0){
-						elastest_is_running = waitElastest()
-						def return_status = this.@ctx.sh script: 'docker ps | grep elastest_', returnStatus:true
-						echo 'elastest_is_running:'+elastest_is_running
-						counter = counter -1
-					}
-
+					def elastest_is_running = startElastest()
 					if (! elastest_is_running){
 						this.@ctx.currentBuild.result = 'FAILURE'
 						return
